# docs/openapi/roadwatch-api.yaml
openapi: 3.0.3
info:
  title: Roadwatch API
  version: 1.0.0
  description: Combined OpenAPI spec for Gateway API (user-facing) and Backend API (auxiliary services). Includes health, auth, complaints, submissions, webhook and analytics endpoints.
servers:
  - url: http://localhost:3100
    description: Local gateway
  - url: http://localhost:4001
    description: Backend auxiliary service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ComplaintCreate:
      type: object
      required: [roadId, title, description, damageType, severity, lat, lng]
      properties:
        roadId:
          type: string
        title:
          type: string
        description:
          type: string
        damageType:
          type: string
        severity:
          type: integer
          minimum: 1
          maximum: 5
        lat:
          type: number
        lng:
          type: number
        capturedLat:
          type: number
        capturedLng:
          type: number
        capturedAt:
          type: string
          format: date-time
        imageCid:
          type: string
    ComplaintSummary:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        severity:
          type: integer
        reportCount:
          type: integer
        lat:
          type: number
        lng:
          type: number
...
paths:
...
  /complaints:
    get:
      summary: List complaints with filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: bounds
          schema:
            type: string
        - in: query
          name: severity
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: Complaint list
          content:
            application/json:
              schema:
                type: object
                properties:
                  complaints:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComplaintSummary'
                  pagination:
                    type: object
    post:
      summary: Create or merge complaint (gateway)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplaintCreate'
      responses:
        '201':
          description: Created or merged
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  message:
                    type: string
                  complaint:
                    $ref: '#/components/schemas/ComplaintSummary'
  /complaints/{id}:
    get:
      summary: Get complaint by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Complaint
        '404':
          description: Not found
...